# Handle underscore similarly to dash