@vvv 2023-05-31
I'll experiment with combining `split_camel_case_bounds` and `split_snake_case_bounds` into a single function...